# Add optimized variant of CMN for HWC to CHW case