… dali/test/python/operator_1/test_crop_mirror_normalize.py